feat(Notifications): ✨dynamically update notification time
